+++ controls-pi/src/main/kotlin/space/kscience/controls/pi/PiPlugin.kt
@@ -2,7 +2,6 @@ package space.kscience.controls.pi
 
 import com.pi4j.Pi4J
 import space.kscience.controls.manager.DeviceManager
-import space.kscience.controls.ports.PortFactory
 import space.kscience.controls.ports.Ports
 import space.kscience.dataforge.context.AbstractPlugin
 import space.kscience.dataforge.context.Context
@@ -10,7 +9,7 @@ import space.kscience.dataforge.context.PluginFactory
 import space.kscience.dataforge.context.PluginTag
 import space.kscience.dataforge.meta.Meta
 import space.kscience.dataforge.names.Name
-import space.kscience.dataforge.names.parseAsName
+import space.kscience.dataforge.names.asName
 import com.pi4j.context.Context as PiContext
 
 public class PiPlugin : AbstractPlugin() {
@@ -22,8 +21,11 @@ public class PiPlugin : AbstractPlugin() {
     public val piContext: PiContext by lazy { createPiContext(context, meta) }
 
     override fun content(target: String): Map<Name, Any> = when (target) {
-        PortFactory.TYPE -> mapOf(
-            PiSerialPort.type.parseAsName() to PiSerialPort,
+        Ports.ASYNCHRONOUS_PORT_TYPE -> mapOf(
+            "serial".asName() to AsynchronousPiPort,
+        )
+        Ports.SYNCHRONOUS_PORT_TYPE -> mapOf(
+            "serial".asName() to SynchronousPiPort,
         )
 
         else -> super.content(target)
@@ -40,6 +42,7 @@ public class PiPlugin : AbstractPlugin() {
 
         override fun build(context: Context, meta: Meta): PiPlugin = PiPlugin()
 
+        @Suppress("UNUSED_PARAMETER")
         public fun createPiContext(context: Context, meta: Meta): PiContext = Pi4J.newAutoContext()
 
     }